+++ muna/beta/metadata/__init__.py
@@ -7,9 +7,10 @@ from ._torch import TorchExporter
 from .coreml import CoreMLInferenceMetadata
 from .executorch import ExecuTorchInferenceBackend, ExecuTorchInferenceMetadata
 from .iree import IREEInferenceBackend, IREEInferenceMetadata
-from .litert import LiteRTInferenceMetadata, LiteRTInterpreterMetadata
-from .onnx import OnnxRuntimeInferenceMetadata
+from .litert import LiteRTInferenceMetadata
+from .onnx import OnnxRuntimeInferenceMetadata, OnnxRuntimeOptimizationLevel
 from .onnxruntime import OnnxRuntimeInferenceSessionMetadata
 from .openvino import OpenVINOInferenceMetadata
 from .qnn import QnnInferenceBackend, QnnInferenceMetadata, QnnInferenceQuantization
 from .tensorrt import CudaArchitecture, TensorRTInferenceMetadata, TensorRTPrecision
+from .tflite import TFLiteInterpreterMetadata