+++ muna/beta/metadata/iree.py
@@ -3,7 +3,7 @@
 #   Copyright © 2025 NatML Inc. All Rights Reserved.
 #
 
-from pydantic import ConfigDict, Field
+from pydantic import Field
 from typing import Literal
 
 from ._torch import PyTorchInferenceMetadataBase
@@ -19,6 +19,7 @@ class IREEInferenceMetadata(PyTorchInferenceMetadataBase):
         model_args (tuple[Tensor,...]): Positional inputs to the model.
         input_shapes (list): Model input tensor shapes. Use this to specify dynamic axes.
         output_keys (list): Model output dictionary keys. Use this if the model returns a dictionary.
+        exporter (TorchExporter): PyTorch exporter to use.
     """
     kind: Literal["meta.inference.iree"] = Field(default="meta.inference.iree", init=False)
     backend: IREEInferenceBackend = Field(
@@ -26,4 +27,3 @@ class IREEInferenceMetadata(PyTorchInferenceMetadataBase):
         description="IREE HAL target backend to execute the model.",
         exclude=True
     )
-    model_config = ConfigDict(arbitrary_types_allowed=True, frozen=True)