+++ muna/beta/metadata/openvino.py
@@ -3,7 +3,7 @@
 #   Copyright © 2025 NatML Inc. All Rights Reserved.
 #
 
-from pydantic import ConfigDict, Field
+from pydantic import Field
 from typing import Literal
 
 from ._torch import PyTorchInferenceMetadataBase
@@ -17,6 +17,6 @@ class OpenVINOInferenceMetadata(PyTorchInferenceMetadataBase):
         model_args (tuple[Tensor,...]): Positional inputs to the model.
         input_shapes (list): Model input tensor shapes. Use this to specify dynamic axes.
         output_keys (list): Model output dictionary keys. Use this if the model returns a dictionary.
+        exporter (TorchExporter): PyTorch exporter to use.
     """
     kind: Literal["meta.inference.openvino"] = Field(default="meta.inference.openvino", init=False)
-    model_config = ConfigDict(arbitrary_types_allowed=True, frozen=True)