+++ muna/beta/metadata/qnn.py
@@ -3,7 +3,7 @@
 #   Copyright © 2025 NatML Inc. All Rights Reserved.
 #
 
-from pydantic import ConfigDict, Field
+from pydantic import Field
 from typing import Literal
 
 from ._torch import PyTorchInferenceMetadataBase
@@ -20,6 +20,7 @@ class QnnInferenceMetadata(PyTorchInferenceMetadataBase):
         model_args (tuple[Tensor,...]): Positional inputs to the model.
         input_shapes (list): Model input tensor shapes. Use this to specify dynamic axes.
         output_keys (list): Model output dictionary keys. Use this if the model returns a dictionary.
+        exporter (TorchExporter): PyTorch exporter to use.
         backend (QnnInferenceBackend): QNN inference backend. Defaults to `cpu`.
         quantization (QnnInferenceQuantization): QNN model quantization mode. This MUST only be specified when backend is `htp`.
     """
@@ -34,4 +35,3 @@ class QnnInferenceMetadata(PyTorchInferenceMetadataBase):
         description="QNN model quantization mode. This MUST only be specified when backend is `htp`.",
         exclude=True
     )
-    model_config = ConfigDict(arbitrary_types_allowed=True, frozen=True)